// src/grpc/proto/proto/main.ts
export interface RunCommandRequest {
  additionalInfo: AdditionalInfo | undefined
  args: string[]
  coreCliHelperId: number
}

export interface RunCommandResponse {}

export interface AdditionalInfo {
  isTerminal: IsTerminal | undefined
  terminalDimensions: TerminalDimensions | undefined
}

export interface IsTerminal {
  stdin: boolean
  stdout: boolean
  stderr: boolean
}

export interface TerminalDimensions {
  width: number
  height: number
}

export interface EchoRequest {
  input: string
}

export interface EchoResponse {
  output: string
}

export interface SendAnalyticsRequest {
  eventName: string
  eventValue: string
}

export interface SendAnalyticsResponse {}
// ...
export const RunCommandRequest = {
  encode(
    message: RunCommandRequest,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.additionalInfo !== undefined) {
      AdditionalInfo.encode(message.additionalInfo, writer.uint32(10).fork()).ldelim()
    }
    for (const v of message.args) {
      writer.uint32(18).string(v!)
    }
    if (message.coreCliHelperId !== 0) {
      writer.uint32(24).uint32(message.coreCliHelperId)
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): RunCommandRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseRunCommandRequest()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break
          }

          message.additionalInfo = AdditionalInfo.decode(reader, reader.uint32())
          continue
        case 2:
          if (tag !== 18) {
            break
          }

          message.args.push(reader.string())
          continue
        case 3:
          if (tag !== 24) {
            break
          }

          message.coreCliHelperId = reader.uint32()
          continue
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(object: any): RunCommandRequest {
    return {
      additionalInfo: isSet(object.additionalInfo)
        ? AdditionalInfo.fromJSON(object.additionalInfo)
        : undefined,
      args: globalThis.Array.isArray(object?.args)
        ? object.args.map((e: any) => globalThis.String(e))
        : [],
      coreCliHelperId: isSet(object.coreCliHelperId)
        ? globalThis.Number(object.coreCliHelperId)
        : 0,
    }
  },

  toJSON(message: RunCommandRequest): unknown {
    const obj: any = {}
    if (message.additionalInfo !== undefined) {
      obj.additionalInfo = AdditionalInfo.toJSON(message.additionalInfo)
    }
    if (message.args?.length) {
      obj.args = message.args
    }
    if (message.coreCliHelperId !== 0) {
      obj.coreCliHelperId = Math.round(message.coreCliHelperId)
    }
    return obj
  },

  create<I extends Exact<DeepPartial<RunCommandRequest>, I>>(
    base?: I,
  ): RunCommandRequest {
    return RunCommandRequest.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<RunCommandRequest>, I>>(
    object: I,
  ): RunCommandRequest {
    const message = createBaseRunCommandRequest()
    message.additionalInfo =
      object.additionalInfo !== undefined && object.additionalInfo !== null
        ? AdditionalInfo.fromPartial(object.additionalInfo)
        : undefined
    message.args = object.args?.map(e => e) || []
    message.coreCliHelperId = object.coreCliHelperId ?? 0
    return message
  },
}

function createBaseRunCommandResponse(): RunCommandResponse {
  return {}
}

export const RunCommandResponse = {
  encode(_: RunCommandResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): RunCommandResponse {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseRunCommandResponse()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(_: any): RunCommandResponse {
    return {}
  },

  toJSON(_: RunCommandResponse): unknown {
    const obj: any = {}
    return obj
  },

  create<I extends Exact<DeepPartial<RunCommandResponse>, I>>(
    base?: I,
  ): RunCommandResponse {
    return RunCommandResponse.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<RunCommandResponse>, I>>(
    _: I,
  ): RunCommandResponse {
    const message = createBaseRunCommandResponse()
    return message
  },
}

function createBaseAdditionalInfo(): AdditionalInfo {
  return { isTerminal: undefined, terminalDimensions: undefined }
}

export const AdditionalInfo = {
  encode(message: AdditionalInfo, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.isTerminal !== undefined) {
      IsTerminal.encode(message.isTerminal, writer.uint32(10).fork()).ldelim()
    }
    if (message.terminalDimensions !== undefined) {
      TerminalDimensions.encode(
        message.terminalDimensions,
        writer.uint32(18).fork(),
      ).ldelim()
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): AdditionalInfo {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseAdditionalInfo()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break
          }

          message.isTerminal = IsTerminal.decode(reader, reader.uint32())
          continue
        case 2:
          if (tag !== 18) {
            break
          }

          message.terminalDimensions = TerminalDimensions.decode(reader, reader.uint32())
          continue
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(object: any): AdditionalInfo {
    return {
      isTerminal: isSet(object.isTerminal)
        ? IsTerminal.fromJSON(object.isTerminal)
        : undefined,
      terminalDimensions: isSet(object.terminalDimensions)
        ? TerminalDimensions.fromJSON(object.terminalDimensions)
        : undefined,
    }
  },

  toJSON(message: AdditionalInfo): unknown {
    const obj: any = {}
    if (message.isTerminal !== undefined) {
      obj.isTerminal = IsTerminal.toJSON(message.isTerminal)
    }
    if (message.terminalDimensions !== undefined) {
      obj.terminalDimensions = TerminalDimensions.toJSON(message.terminalDimensions)
    }
    return obj
  },

  create<I extends Exact<DeepPartial<AdditionalInfo>, I>>(base?: I): AdditionalInfo {
    return AdditionalInfo.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<AdditionalInfo>, I>>(
    object: I,
  ): AdditionalInfo {
    const message = createBaseAdditionalInfo()
    message.isTerminal =
      object.isTerminal !== undefined && object.isTerminal !== null
        ? IsTerminal.fromPartial(object.isTerminal)
        : undefined
    message.terminalDimensions =
      object.terminalDimensions !== undefined && object.terminalDimensions !== null
        ? TerminalDimensions.fromPartial(object.terminalDimensions)
        : undefined
    return message
  },
}
// ...
export type MainService = typeof MainService
export const MainService = {
  runCommand: {
    path: '/proto.Main/RunCommand',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: RunCommandRequest) =>
      Buffer.from(RunCommandRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => RunCommandRequest.decode(value),
    responseSerialize: (value: RunCommandResponse) =>
      Buffer.from(RunCommandResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => RunCommandResponse.decode(value),
  },
} as const

export interface MainServer extends UntypedServiceImplementation {
  runCommand: handleUnaryCall<RunCommandRequest, RunCommandResponse>
}

export type CoreCLIHelperService = typeof CoreCLIHelperService
export const CoreCLIHelperService = {
  echo: {
    path: '/proto.CoreCLIHelper/Echo',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: EchoRequest) =>
      Buffer.from(EchoRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => EchoRequest.decode(value),
    responseSerialize: (value: EchoResponse) =>
      Buffer.from(EchoResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => EchoResponse.decode(value),
  },
  sendAnalytics: {
    path: '/proto.CoreCLIHelper/SendAnalytics',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: SendAnalyticsRequest) =>
      Buffer.from(SendAnalyticsRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => SendAnalyticsRequest.decode(value),
    responseSerialize: (value: SendAnalyticsResponse) =>
      Buffer.from(SendAnalyticsResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => SendAnalyticsResponse.decode(value),
  },
  keychainGetPassword: {
    path: '/proto.CoreCLIHelper/KeychainGetPassword',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: KeychainGetPasswordRequest) =>
      Buffer.from(KeychainGetPasswordRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => KeychainGetPasswordRequest.decode(value),
    responseSerialize: (value: KeychainGetPasswordResponse) =>
      Buffer.from(KeychainGetPasswordResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => KeychainGetPasswordResponse.decode(value),
  },
  keychainSetPassword: {
    path: '/proto.CoreCLIHelper/KeychainSetPassword',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: KeychainSetPasswordRequest) =>
      Buffer.from(KeychainSetPasswordRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => KeychainSetPasswordRequest.decode(value),
    responseSerialize: (value: KeychainSetPasswordResponse) =>
      Buffer.from(KeychainSetPasswordResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => KeychainSetPasswordResponse.decode(value),
  },
  keychainDeletePassword: {
    path: '/proto.CoreCLIHelper/KeychainDeletePassword',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: KeychainDeletePasswordRequest) =>
      Buffer.from(KeychainDeletePasswordRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => KeychainDeletePasswordRequest.decode(value),
    responseSerialize: (value: KeychainDeletePasswordResponse) =>
      Buffer.from(KeychainDeletePasswordResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => KeychainDeletePasswordResponse.decode(value),
  },
  keychainFindCredentials: {
    path: '/proto.CoreCLIHelper/KeychainFindCredentials',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: KeychainFindCredentialsRequest) =>
      Buffer.from(KeychainFindCredentialsRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => KeychainFindCredentialsRequest.decode(value),
    responseSerialize: (value: KeychainFindCredentialsResponse) =>
      Buffer.from(KeychainFindCredentialsResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => KeychainFindCredentialsResponse.decode(value),
  },
  runPeerPlugin: {
    path: '/proto.CoreCLIHelper/RunPeerPlugin',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: RunPeerPluginRequest) =>
      Buffer.from(RunPeerPluginRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => RunPeerPluginRequest.decode(value),
    responseSerialize: (value: RunPeerPluginResponse) =>
      Buffer.from(RunPeerPluginResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => RunPeerPluginResponse.decode(value),
  },
} as const
// ...
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never
    }